greedy monocarp 800

+++ codeforces/a/prep_for_exam.cpp
@@ -15,10 +15,10 @@
 #include <map>
 #include <set>
 using namespace std;
-using i32 = int32_t;
-using u32 = uint32_t;
-using i64 = int64_t;
-using u64 = uint64_t;
+using i32 = int;
+using u32 = unsigned int;
+using i64 = long long;
+using u64 = unsigned long long;
 using f32 = float;
 using f64 = double;
 using str = string;
